REACT_APP_X for common file data
