+++ .env
@@ -1,2 +1,9 @@
+REACT_APP_NODE=localhost
+REACT_APP_PORT=3000
+
 PUBLIC_URL = /public
+
+REACT_APP_DOMAIN=DiagnosticImaging
+
+REACT_APP_MANUFACTURER=Illuminati
